requirements.txt
Django>=4.2
gunicorn
whitenoise
dj-database-url
python-dotenv
Pillow
psycopg[binary]      # <-- agrega esto (o psycopg2-binary)      # solo si lo usás en tu PC (podés dejarlo)
django-cloudinary-storage
cloudinary
django-cloudinary-storage>=0.3.0
cloudinary>=1.41
